--- crates/cubecl-linalg/src/matmul/components/global/loader/mod.rs
@@ -1,2 +1,0 @@
-pub mod r#async;
-pub mod sync;
