--- crates/cubecl-linalg/src/matmul/components/global/loader/sync/mod.rs
@@ -1,5 +1,0 @@
-mod buffer;
-mod full;
-
-pub use buffer::*;
-pub use full::*;
