+++ crates/cubecl-linalg/src/matmul/components/global/mod.rs
@@ -1,5 +1,4 @@
 pub mod args;
-pub mod loader;
 pub mod multi_stage;
 pub mod single_stage;
 pub mod tensor_view;
