+++ crates/cubecl-linalg/src/matmul/components/global/multi_stage/loading/mod.rs
@@ -1,0 +1,5 @@
+mod r#async;
+mod sync;
+
+pub use r#async::*;
+pub use sync::*;
