+++ crates/cubecl-linalg/src/matmul/components/global/multi_stage/mod.rs
@@ -1,2 +1,7 @@
+mod loading;
+
 pub mod double_buffering;
 pub mod specialized;
+
+pub use loading::*;
+
