+++ crates/cubecl-linalg/src/matmul/components/global/single_stage/loading/async/base.rs
