+++ crates/cubecl-linalg/src/matmul/components/global/single_stage/loading/async/maximize_slice_length.rs
