+++ crates/cubecl-linalg/src/matmul/components/global/single_stage/loading/async/maximize_unit_count.rs
