+++ crates/cubecl-linalg/src/matmul/components/global/single_stage/loading/mod.rs
@@ -1,0 +1,5 @@
+mod sync;
+mod r#async;
+
+pub use sync::*;
+pub use r#async::*;
