+++ crates/cubecl-linalg/src/matmul/components/global/single_stage/loading/sync/base.rs
