+++ crates/cubecl-linalg/src/matmul/components/global/single_stage/mod.rs
@@ -1,5 +1,7 @@
 pub mod simple;
 
 mod config;
+mod loading;
 
 pub use config::*;
+pub use loading::*;
