+++ crates/cubecl-linalg/src/matmul/components/global/single_stage/simple/matmul.rs
@@ -1,9 +1,7 @@
 use crate::matmul::components::global::base::InputLoader;
-use crate::matmul::components::global::loader::sync::{
-    SyncFullLhsLoader, SyncFullLoadingStrategy, SyncFullRhsLoader,
-};
 use crate::matmul::components::global::output_loader::Unloader;
 use crate::matmul::components::global::single_stage::Config;
+use crate::matmul::components::global::single_stage::SyncFullLoadingStrategy;
 use crate::matmul::components::global::ZeroAccumulatorLoader;
 use crate::matmul::components::global::{GlobalMatmul, SyncInputLoader};
 use crate::matmul::components::stage::multi_buffer::{LhsReader, RhsReader};
