+++ crates/cubecl-linalg/src/matmul/components/global/single_stage/simple/matmul_barrier.rs
@@ -1,9 +1,8 @@
 use crate::matmul::components::global::base::AsyncInputLoader;
 use crate::matmul::components::global::base::InputLoader;
-use crate::matmul::components::global::loader::r#async::AsyncFullLoadingStrategy;
-use crate::matmul::components::global::loader::r#async::{AsyncLhsLoader, AsyncRhsLoader};
 use crate::matmul::components::global::output_loader::Unloader;
 use crate::matmul::components::global::single_stage::Config;
+use crate::matmul::components::global::single_stage::AsyncFullLoadingStrategy;
 use crate::matmul::components::global::GlobalMatmul;
 use crate::matmul::components::global::ZeroAccumulatorLoader;
 use crate::matmul::components::stage::multi_buffer::{LhsReader, RhsReader};
