+++ crates/cubecl-linalg/src/matmul/kernels/matmul/algorithm/simple.rs
@@ -6,7 +6,7 @@ use crate::matmul::components::{
     batch::{self, CubeCountDispatch, CubeDispatch},
     global::{
         self,
-        loader::sync::{CyclicCoalescedLoading, SyncFullLoadingStrategy},
+        single_stage::{CyclicCoalescedLoading, SyncFullLoadingStrategy},
     },
     stage::{self, ColMajorTilingOrder, RowMajorTilingOrder},
     tile, MatmulProblem, MatmulSelection,
