+++ crates/cubecl-linalg/src/matmul/kernels/matmul/algorithm/simple_pipelined.rs
@@ -4,7 +4,7 @@ use std::marker::PhantomData;
 
 use crate::matmul::components::{
     batch::{self, CubeCountDispatch, CubeDispatch},
-    global::{self, loader::r#async::CyclicWindowLoading},
+    global::{self, single_stage::CyclicWindowLoading},
     stage::{self, ColMajorTilingOrder, RowMajorTilingOrder},
     tile, MatmulProblem, MatmulSelection,
 };
